Deprecate DefaultScheduler, make AReal ~degenerate to it
