+++ src/prime_rl/orchestrator/utils.py
@@ -8,6 +8,7 @@ from openai.types.completion_usage import CompletionUsage
 from rich.console import Console
 from rich.table import Table
 
+from prime_rl.orchestrator.config import SamplingConfig
 from prime_rl.utils.utils import (
     format_num,
     format_time,
@@ -156,3 +157,19 @@ def print_benchmark(history: dict[str, list[Any]]) -> None:
 
     # Display table
     console.print(table)
+
+
+def prepare_sampling_args(sampling_config: SamplingConfig) -> dict:
+    sampling_args = dict(sampling_config)
+    # Convert SamplingConfig to vLLM OAI sampling args
+    # https://docs.vllm.ai/en/latest/serving/openai_compatible_server.html#extra-parameters_2
+    sampling_args["top_p"] = 1.0
+    sampling_args["logprobs"] = True
+    sampling_args["extra_body"] = {
+        "return_tokens_as_token_ids": True,
+        "top_k": -1,
+        "min_p": 0.0,
+    }
+    sampling_args["extra_body"]["min_tokens"] = sampling_args.pop("min_tokens")
+    sampling_args["extra_body"]["repetition_penalty"] = sampling_args.pop("repetition_penalty")
+    return sampling_args
